refactor: organise workflow helpers in folders
