+++ nodes/web/helpers/workflow/form/main.js
@@ -1,0 +1,4 @@
+import { showForm, showWorkflowsList } from "./style.js";
+import { populateSubmitForm, saveWorkflow } from "./utils.js";
+
+export { populateSubmitForm, saveWorkflow, showForm, showWorkflowsList };
