--- nodes/web/helpers/workflow/main.js
@@ -1,23 +1,0 @@
-import { saveWorkflow, showWorkflowsList } from "./form.js";
-import { showNodeOrderEditor } from "./nodeOrder.js";
-import {
-  $el,
-  cleanLocalStorage,
-  createMenuItem,
-  deleteWorkflowFromStorage,
-  findMenuList,
-  requiresAuth,
-  showMessage,
-} from "./utils.js";
-export {
-  $el,
-  cleanLocalStorage,
-  createMenuItem,
-  deleteWorkflowFromStorage,
-  findMenuList,
-  requiresAuth,
-  saveWorkflow,
-  showMessage,
-  showNodeOrderEditor,
-  showWorkflowsList,
-};
