+++ nodes/web/helpers/workflow/node_order/main.js
@@ -1,0 +1,3 @@
+import { showNodeOrderEditor } from "./style.js";
+
+export { showNodeOrderEditor };
