+++ nodes/web/signature_api/workflow_services.js
@@ -1,6 +1,6 @@
 import { getAccessToken, signatureApiBaseUrl } from "./main.js";
 
-export const getWorkflowsListForForm = async (element, offset = 0, limit = 100, search = null) => {
+const getWorkflowsListForForm = async (element, offset = 0, limit = 100, search = null) => {
   const accessToken = getAccessToken();
   let url = `${signatureApiBaseUrl}/api/v1_management/workflow?offset=${offset}&limit=${limit}`;
   if (search) {
@@ -28,7 +28,7 @@ export const getWorkflowsListForForm = async (element, offset = 0, limit = 100, 
   }
 };
 
-export const getWorkflowById = async (workflowId) => {
+const getWorkflowById = async (workflowId) => {
   const accessToken = getAccessToken();
   const url = `${signatureApiBaseUrl}/api/v1_management/workflow/${workflowId}`;
   const response = await fetch(url, {
@@ -47,7 +47,7 @@ export const getWorkflowById = async (workflowId) => {
   return parsedResponse.result;
 };
 
-export const getWorkflowVersions = async (workflowId, offset = 0, limit = 100) => {
+const getWorkflowVersions = async (workflowId, offset = 0, limit = 100) => {
   const accessToken = getAccessToken();
   const url = `${signatureApiBaseUrl}/api/v1_management/workflow/${workflowId}/version?includeWorkflow=true&offset=${offset}&limit=${limit}`;
   const response = await fetch(url, {
@@ -65,3 +65,33 @@ export const getWorkflowVersions = async (workflowId, offset = 0, limit = 100) =
   const parsedResponse = await response.json();
   return parsedResponse.result;
 };
+
+const getManifest = async (workflow) => {
+  try {
+    const url = window.location.href + "flow/create_manifest";
+
+    const response = await fetch(url, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({
+        workflow: workflow,
+      }),
+    });
+
+    if (!response.ok) {
+      const errorText = await response.text();
+      console.error("Error getting manifest:", errorText);
+      throw new Error(`Failed to get manifest: ${response.status} ${response.statusText}`);
+    }
+
+    const data = await response.json();
+    return data;
+  } catch (error) {
+    console.error("Error in getManifest:", error);
+    throw error;
+  }
+};
+
+export { getManifest, getWorkflowById, getWorkflowsListForForm, getWorkflowVersions };
